requirements.txt
Flask
opencv-python
numpy
Pillow
pytesseract
